+++ packages/terra-form/tests/nightwatch/FormTestRoutes.jsx
@@ -21,9 +21,11 @@ import PopulatedTextarea from './textarea/PopulatedTextarea';
 import DefaultTextareaField from './textarea-field/DefaultTextareaField';
 import PopulatedTextareaField from './textarea-field/PopulatedTextareaField';
 import DefaultSelect from './select/DefaultSelect';
+import ChoicesPropSelect from './select/ChoicesPropSelect';
 import PopulatedSelect from './select/PopulatedSelect';
 import PopulatedSelectControlled from './select/PopulatedSelectControlled';
 import DefaultSelectField from './select-field/DefaultSelectField';
+import ChoicesPropSelectField from './select-field/ChoicesPropSelectField';
 import PopulatedSelectField from './select-field/PopulatedSelectField';
 
 const routes = (
@@ -46,9 +48,11 @@ const routes = (
     <Route path="/tests/form-tests/textarea-field/default" component={DefaultTextareaField} />
     <Route path="/tests/form-tests/textarea-field/populated" component={PopulatedTextareaField} />
     <Route path="/tests/form-tests/select/default" component={DefaultSelect} />
+    <Route path="/tests/form-tests/select/choices_prop" component={ChoicesPropSelect} />
     <Route path="/tests/form-tests/select/populated" component={PopulatedSelect} />
     <Route path="/tests/form-tests/select/populated_controlled" component={PopulatedSelectControlled} />
     <Route path="/tests/form-tests/select-field/default" component={DefaultSelectField} />
+    <Route path="/tests/form-tests/select-field/choices_prop" component={ChoicesPropSelectField} />
     <Route path="/tests/form-tests/select-field/populated" component={PopulatedSelectField} />
   </div>
 );